Fixed a lot of things to make it work with a database
Made it so the API will send a request to the database and it will send back a response according to what data was POSTed. It's possible to get your horoscope sign now depending on the date selected. The delete button also works, it destroys the session. Code clean up and minor css changes